(fix)Make proper __init__ structure
